Remove Pneumatics Code
At this point it seems pretty unlikely that we will be using pneumatics on ERSO, so I removed all of the pneumatics-related code. If we end up using a different type of mechansim to deploy/raise the intake (motor with ratchet) or adjust the shooter hood (servo motor) then we can add those subsystems back in, but that also seems pretty unlikely.